typescript refactoring of deployments
