--- widgets/deployments/src/props/DeploymentsViewDefaultProps.ts
@@ -1,3 +1,0 @@
-export default {
-    allowedSettingTo: ['tenant', 'global']
-};
